Feature: Private Chat
User can now send private messages to the online users. The private messages appear in the same space as the free node messages but with a different color. The user can now switch between private and public chat

Known Bug: The private messages are not shown to the sender but only the receiver. There is yet no way to keep track of what what sent in a private chat but only what was recieved.